change Summ(in Order) - from int to decimal

+++ DogSitterMarketplaceApi/Models/OrdersDto/Request/OrderUpdateDto.cs
@@ -10,7 +10,7 @@
 
         public int SitterWorkId { get; set; }
 
-        public int Summ { get; set; }
+        public decimal Summ { get; set; }
 
         public DateTime DateStart { get; set; }
