+++ DogSitterMarketplaceBll/Models/Orders/Request/OrderCreateRequest.cs
@@ -8,7 +8,7 @@
 
         public int SitterWorkId { get; set; }
 
-        public int Summ { get; set; }
+        public decimal Summ { get; set; }
 
         public DateTime DateStart { get; set; }
